+++ packages/terra-avatar/src/terra-dev-site/doc/example/AvatarTwoInitials.jsx
@@ -1,9 +1,0 @@
-import React from 'react';
-// eslint-disable-next-line import/no-extraneous-dependencies, import/no-unresolved, import/extensions
-import Avatar from 'terra-avatar/lib/Avatar';
-
-const AvatarTwoInitials = () => (
-  <Avatar ariaLabel="John Doe" initials="JS" />
-);
-
-export default AvatarTwoInitials;