+++ packages/terra-button/src/terra-dev-site/test/button/Variants/ActionButton.test.jsx
@@ -7,6 +7,7 @@ const ActionButton = () => (
   <div id="action" style={{ padding: '5px' }}>
     <Button id="actionButton" text="Action" icon={<IconSquare />} variant="action" />
     <Button id="actionButtonDisabled" text="Action Disabled" icon={<IconSquare />} variant="action" isDisabled />
-  </div>);
+  </div>
+);
 
 export default ActionButton;