+++ packages/terra-form-input/src/terra-dev-site/doc/example/uncontrolled/DefaultInvalidExample.jsx
@@ -1,16 +1,18 @@
 import React from 'react';
-
+import Field from 'terra-form-field';
 // eslint-disable-next-line import/no-extraneous-dependencies, import/no-unresolved, import/extensions
 import Input from 'terra-form-input';
 
 const DefaultInvalidExample = () => (
   <div>
-    <h4>Default</h4>
-    <Input
-      name="default error input"
-      defaultValue="Default Error Input – Uncontrolled"
-      isInvalid
-    />
+    <Field label="Default" htmlFor="uncontrolled-default-invalid">
+      <Input
+        name="default error input"
+        id="uncontrolled-default-invalid"
+        defaultValue="Default Error Input – Uncontrolled"
+        isInvalid
+      />
+    </Field>
   </div>
 );