+++ packages/terra-progress-bar/CHANGELOG.md
@@ -4,6 +4,26 @@ Changelog
 Unreleased
 ----------
 
+3.2.0 - (January 22, 2019)
+------------------
+### Changed
+* Minor dependency version bump
+
+3.1.0 - (January 22, 2019)
+------------------
+### Changed
+* Minor dependency version bump
+
+3.0.0 - (January 8, 2019)
+------------------
+### Changed
+* Updated theme variables
+
+2.31.0 - (January 2, 2019)
+------------------
+### Changed
+* Minor dependency version bump
+
 2.30.0 - (November 19, 2018)
 ------------------
 ### Changed